Counter example now uses SimpleStorage and GraphQL
