--- linera-examples/counter-graphql/Cargo.toml
@@ -1,29 +1,0 @@
-[package]
-name = "counter-graphql"
-version = "0.1.0"
-authors = ["Linera <contact@linera.io>"]
-edition = "2021"
-
-[dependencies]
-async-graphql = { version = "5.0.7", default-features = false }
-async-trait = "0.1.52"
-bcs = "0.1.3"
-futures = "0.3.17"
-hex = "0.4.3"
-linera-sdk = { path = "../../linera-sdk" }
-linera-views = { path = "../../linera-views" }
-serde = { version = "1.0.152", features = ["derive"] }
-serde_json = "1.0.93"
-thiserror = "1.0.31"
-
-[dev-dependencies]
-linera-sdk = { path = "../../linera-sdk", features = ["test"] }
-webassembly-test = "0.1.0"
-
-[[bin]]
-name = "counter_graphql_contract"
-path = "src/contract.rs"
-
-[[bin]]
-name = "counter_graphql_service"
-path = "src/service.rs"
