+++ linera-examples/counter/Cargo.toml
@@ -5,12 +5,14 @@ authors = ["Linera <contact@linera.io>"]
 edition = "2021"
 
 [dependencies]
+async-graphql = { version = "5.0.7", default-features = false }
 async-trait = "0.1.52"
 bcs = "0.1.3"
 futures = "0.3.17"
 linera-sdk = { path = "../../linera-sdk" }
 log = "0.4.17"
 serde = { version = "1.0.130", features = ["derive"] }
+serde_json = "1.0.96"
 thiserror = "1.0.31"
 
 [dev-dependencies]
