+++ linera-examples/counter/web-frontend/.gitignore
