+++ linera-examples/counter/web-frontend/.prettierignore
