+++ linera-examples/counter/web-frontend/public/favicon.ico
