+++ linera-examples/counter/web-frontend/public/index.html
