+++ linera-examples/counter/web-frontend/src/GraphQLProvider.js
