+++ linera-examples/counter/web-frontend/src/index.js
