+++ linera-examples/counter/web-frontend/tailwind.config.js
